# inventory-api/docs/openapi/runtime-baseline.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "inventory-api runtime baseline",
    "version": "0.5.0",
    "description": "Baseline OpenAPI parcial y factual para superficies críticas verificadas en runtime. No cubre todo el monolito y se complementa con un manifiesto explícito de exclusiones contractuales."
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/orders": {
      "get": {
        "operationId": "listOrders",
        "summary": "Listar pedidos visibles para roles operativos habilitados",
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/orders",
          "routeFile": "src/routes/order.routes.js",
          "expressPath": "/"
        },
        "responses": {
          "200": {
            "description": "Listado de pedidos"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      },
      "post": {
        "operationId": "createOrder",
        "summary": "Registrar pedido de ventas",
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/orders",
          "routeFile": "src/routes/order.routes.js",
          "expressPath": "/"
        },
        "responses": {
          "201": {
            "description": "Pedido creado"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      }
    },
...
    "/api/sales-routes/company/agents/{userId}/goals": {
      "put": {
        "operationId": "saveCompanySalesRouteAgentGoals",
        "summary": "Guardar metas de agente para rutas comerciales",
        "tags": [
          "sales-routes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/sales-routes",
          "routeFile": "src/routes/sales-route.routes.js",
          "expressPath": "/company/agents/:userId/goals"
        },
        "responses": {
          "200": {
            "description": "Metas actualizadas"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      }
    },
...
    "/api/agent/dashboard": {
      "get": {
        "operationId": "getAgentDashboard",
        "summary": "Consultar dashboard del workspace de agente",
        "tags": [
          "agent-workspace"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/agent",
          "routeFile": "src/routes/agent.routes.js",
          "expressPath": "/dashboard"
        },
        "responses": {
          "200": {
            "description": "Dashboard de agente"
          },
          "401": {
            "description": "Token faltante o inválido"
          }
        }
      }
    },
    "/api/agent/stores": {
      "get": {
        "operationId": "listAgentStores",
        "summary": "Listar tiendas visibles en el workspace del agente",
        "tags": [
          "agent-workspace"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/agent",
          "routeFile": "src/routes/agent.routes.js",
          "expressPath": "/stores"
        },
        "responses": {
          "200": {
            "description": "Listado de tiendas"
          },
          "401": {
            "description": "Token faltante o inválido"
          }
        }
      }
    },
    "/api/agent/stores/{storeId}": {
      "get": {
        "operationId": "getAgentStoreDetail",
        "summary": "Consultar detalle de tienda en workspace de agente",
        "tags": [
          "agent-workspace"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/agent",
          "routeFile": "src/routes/agent.routes.js",
          "expressPath": "/stores/:storeId"
        },
        "responses": {
          "200": {
            "description": "Detalle de tienda"
          },
          "404": {
            "description": "Tienda no encontrada"
          }
        }
      }
    },
    "/api/agent/stores/{storeId}/order-context": {
      "get": {
        "operationId": "getAgentStoreOrderContext",
        "summary": "Consultar contexto de pedido de una tienda",
        "tags": [
          "agent-workspace"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/agent",
          "routeFile": "src/routes/agent.routes.js",
          "expressPath": "/stores/:storeId/order-context"
        },
        "responses": {
          "200": {
            "description": "Contexto de pedido"
          },
          "404": {
            "description": "Tienda no encontrada"
          }
        }
      }
    },
    "/api/agent/stores/{storeId}/orders": {
      "post": {
        "operationId": "createAgentStoreOrder",
        "summary": "Crear pedido de agente para una tienda",
        "tags": [
          "agent-workspace"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/agent",
          "routeFile": "src/routes/agent.routes.js",
          "expressPath": "/stores/:storeId/orders"
        },
        "responses": {
          "201": {
            "description": "Pedido creado"
          },
          "401": {
            "description": "Token faltante o inválido"
          },
          "403": {
            "description": "Fuera del tenant o permisos insuficientes"
          }
        }
      }
    },
    "/api/agent/visits": {
      "get": {
        "operationId": "listAgentVisits",
        "summary": "Listar visitas registradas del agente",
        "tags": [
          "agent-workspace"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/agent",
          "routeFile": "src/routes/agent.routes.js",
          "expressPath": "/visits"
        },
        "responses": {
          "200": {
            "description": "Listado de visitas"
          },
          "401": {
            "description": "Token faltante o inválido"
          }
        }
      },
      "post": {
        "operationId": "createAgentVisit",
        "summary": "Registrar visita del agente",
        "tags": [
          "agent-workspace"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-runtime-source": {
          "mountPath": "/api/agent",
          "routeFile": "src/routes/agent.routes.js",
          "expressPath": "/visits"
        },
        "responses": {
          "201": {
            "description": "Visita registrada"
          },
          "401": {
            "description": "Token faltante o inválido"
          }
        }
      }
    },
...
  }
}